Component: generalize __deepcopy__ method for use elsewhere
